moves redirection to processwrapper instead of processprovider
